test(mtp): enroll grammar transactions in Apple coverage
